backend/requirements.txt
Django==5.2.1
djangorestframework
Pillow
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary